Move uuid tests into tests/test_strings.rs
Drop the tests/hypothesis/ directory entirely; uuids() returns a string,
so its tests fit naturally alongside the other formatted-string tests
(ip_addresses, characters, text, regex).
