--- tests/hypothesis/main.rs
@@ -1,7 +1,0 @@
-//! Tests ported from hypothesis-python/tests/. Add one `mod <name>;`
-//! per ported file, alphabetical.
-
-#[path = "../common/mod.rs"]
-mod common;
-
-mod uuids;
